Consolidate more functions into util.py.
