+++ lib/build.py
@@ -151,49 +151,6 @@ def validate_memory_quota(context):
 		print >> sys.stderr, 'For a total of:', required, 'MB'
 		sys.exit(1)
 
-# FIXME this wants to be a dict with 'typename' as the key.
-package_types = [
-	# A + at the start of the job type indicates it is a post-deploy errand
-	# A - at the start of the job type indicates it is a pre-delete errand
-	{
-		'typename': 'app',
-		'flags': [ 'requires_cf_cli', 'is_app' ],
-	},
-	{
-		'typename': 'app-broker',
-		'flags': [ 'requires_cf_cli', 'is_app', 'is_broker', 'is_broker_app' ],
-	},
-	{
-		'typename': 'external-broker',
-		'flags': [ 'requires_cf_cli', 'is_broker', 'is_external_broker' ],
-	},
-	{
-		'typename': 'buildpack',
-		'flags': [ 'requires_cf_cli', 'is_buildpack' ],
-	},
-	{
-		'typename': 'docker-bosh',
-		'flags': [ 'requires_docker_bosh', 'is_docker_bosh', 'is_docker' ],
-		'jobs':  [ 'docker-bosh' ],
-	},
-	{
-		'typename': 'docker-app',
-		'flags': [ 'requires_cf_cli', 'is_app', 'is_docker_app', 'is_docker' ],
-	},
-	{
-		'typename': 'docker-app-broker',
-		'flags': [ 'requires_cf_cli', 'is_app', 'is_broker', 'is_broker_app', 'is_docker_app', 'is_docker' ],
-	},
-	{
-		'typename': 'blob',
-		'flags': [ 'is_blob' ],
-	},
-	{
-		'typename': 'bosh-release',
-		'flags': [ 'is_bosh_release' ],
-	}
-]
-
 def create_bosh_release(context):
 	target = os.getcwd()
 	bosh('init', 'release')
@@ -417,21 +374,6 @@ def create_tile(context):
 	print
 	print 'created tile', pivotal_file
 
-def download_docker_release():
-	release_name = 'docker'
-	release_version = DOCKER_BOSHRELEASE_VERSION
-	release_file = release_name + '-boshrelease-' + release_version + '.tgz'
-	release_tarball = release_file
-	if not os.path.isfile(release_tarball):
-		url = 'https://bosh.io/d/github.com/cf-platform-eng/docker-boshrelease?v=' + release_version
-		download(url, release_tarball)
-	return {
-		'tarball': release_tarball,
-		'name': release_name,
-		'version': release_version,
-		'file': release_file,
-	}
-
 def add_bosh_release(context, package):
 	with cd('..'):
 		tarball = os.path.realpath(package['path'])
@@ -450,41 +392,6 @@ def add_bosh_release(context, package):
 			'version': release_version,
 		}
 	]
-
-def download_docker_image(docker_image, target_file, cache=None):
-	try:
-		from docker.client import Client
-		from docker.utils import kwargs_from_env
-		kwargs = kwargs_from_env()
-		kwargs['tls'] = False
-		docker_cli = Client(**kwargs)
-		image = docker_cli.get_image(docker_image)
-		image_tar = open(target_file,'w')
-		image_tar.write(image.data)
-		image_tar.close()
-	except Exception as e:
-		if cache is not None:
-			cached_file = os.path.join(cache, docker_image.lower().replace('/','-').replace(':','-') + '.tgz')
-			if os.path.isfile(cached_file):
-				print 'using cached version of', docker_image
-				urllib.urlretrieve(cached_file, target_file)
-				return
-			print >> sys.stderr, docker_image, 'not found in cache', cache
-			sys.exit(1)
-		if isinstance(e, KeyError):
-			print >> sys.stderr, 'docker not configured on this machine (or environment variables are not properly set)'
-		else:
-			print >> sys.stderr, docker_image, 'not found on local machine'
-			print >> sys.stderr, 'you must either pull the image, or download it and use the --docker-cache option'
-		sys.exit(1)
-
-def bosh_extract(output, properties):
-	result = {}
-	for l in output.split('\n'):
-		for p in properties:
-			if l.startswith(p['pattern']):
-				result[p['label']] = l.split(':', 1)[-1].strip()
-	return result
 
 def bosh(*argv):
 	argv = list(argv)
